requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.0.10
pypdf
faiss-cpu
google-generativeai
python-dotenv
